Added basic testing suite
Right now it just shallow renders all of the Components, more specific tests still
need written for each component.

There was a myriad of problems encountered when trying to test crypto.js. In order
to test the private functions I will need to get rewired.js or babel-plugin-rewire
working.

Additionally the AES library that I am using is thrown off by JSDOM. It tries to
determine if it is in NodeJS or a browser by checking if window is undefined. JSDOM
defines a window object, but it does not implement the WebCrypt API, so when the
AES library attempts to encrypt data it throws an error that the enviornment is
unsupported.

This could in theory be fixed by putting Jest into NodeJS mode, but that would break
all of the React Component tests.